fixes exposed calculate order code, though perhaps now doubled in a way we could avoid. Also dynamically passes orderID to confirmation page (though I think we'll eventually want to set a different value for this).
